# docs/openapi.yaml
openapi: 3.0.0
info:
  title: View Counter API
  description: |
    A privacy-first, lightweight analytics backend for tracking views and events across multiple applications.
    
    ### Key Features:
    - **Self-managed**: Auto-creates database and tables if needed.
    - **Privacy-First**: No cookies, lightweight tracking.
    - **Secure**: Helmet, rate-limiting, and SQL injection protection.
    - **Rich Analytics**: Page tracking, referrer analysis, user agent parsing.

    ### Authentication
    The tracking endpoints (`/registerView`, `/event`) are public — a browser on
    your site must be able to reach them. Every **analytics read** endpoint
    requires an API key in the `x-api-key` header. Configure keys via the
    `READ_API_KEYS` environment variable; with none set the read API returns
    503 rather than serving data unauthenticated.

    Keys can be **scoped** to specific apps via `apiKeys` in `allowed.json`. A
    key used against an app outside its scope receives 403, and `GET /apps`
    lists only the apps that key may read.

    Provisioning (`POST /apps`) uses a separate **admin** credential
    (`ADMIN_API_KEYS`): a read key cannot provision, and an admin key cannot
    read analytics.
  version: 3.0.0
  contact:
    name: Harsh Ankur
    url: https://github.com/harshankur/viewcounter
servers:
  - url: http://localhost:3030
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Read-API credential, compared in constant time. Minimum 32 characters.
        Several may be configured so one consumer can be revoked independently.

paths:
  /registerView:
    get:
      summary: Register a page view
      description: Track a visitor landing on a page. Automatically parses the User-Agent header.
      parameters:
        - name: appId
          in: query
          required: true
          schema:
            type: string
          description: The registered application ID.
        - name: deviceSize
          in: query
          required: true
          schema:
            type: string
            enum: [small, medium, large]
          description: Screen size category of the user device.
        - name: page
          in: query
          schema:
            type: string
          description: Path of the page (e.g., /blog/post-1).
        - name: title
          in: query
          schema:
            type: string
          description: Title of the page.
        - name: referrer
          in: query
          schema:
            type: string
          description: >-
            Where the visitor came from, normally `document.referrer`. Absent or
            empty records a direct visit. The request's own Referer header is
            never used, since from a browser it names the tracked page itself.
        - name: sessionId
          in: query
          schema:
            type: string
          description: Unique identifier for the user session.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success!
                  duplicate:
                    type: boolean
                    example: false
        '400':
          description: Invalid Parameters
        '422':
          description: Validation Failed
        '429':
          description: Too Many Requests

  /event:
    post:
      summary: Track a custom event
      description: Track arbitrary user interactions (clicks, form submits, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appId, eventType]
              properties:
                appId:
                  type: string
                eventType:
                  type: string
                  example: button_click
                eventData:
                  type: object
                  description: Custom JSON metadata for the event.
                sessionId:
                  type: string
      responses:
        '200':
          description: Event tracked
        '400':
          description: Invalid appId

  /stats/{appId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Get aggregated statistics
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalViews:
                    type: integer
                  uniqueViews:
                    type: integer
                  uniqueVisitors:
                    type: integer
                  last24Hours:
                    type: integer
                  byCountry:
                    type: array
                  byDevice:
                    type: array
        '400':
          description: Invalid appId

  /trends/{appId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Get time-based trends
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [hourly, daily, weekly]
            default: daily
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Success

  /referrers/{appId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Get top traffic sources
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success

  /browsers/{appId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Get browser/OS breakdown
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success

  /pages/{appId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Get most viewed pages
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success

  /apps:
    get:
      summary: List apps visible to the presented key
      description: Returns only the apps within the presented key's scope.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Apps in scope
        '401':
          description: Missing or invalid API key
    post:
      summary: Provision a new app
      description: |
        Creates the app's table, records it in the registry, and adds it to the
        live allowlist without a restart. Requires an admin key. Idempotent.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [appId]
              properties:
                appId:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{1,64}$'
                  description: Becomes a table name. May not start with an underscore.
                  example: newcustomer
                origins:
                  type: array
                  maxItems: 20
                  items:
                    type: string
                    format: uri
                  description: Site origins permitted to write to this app.
                  example: ['https://newcustomer.example']
      responses:
        '200':
          description: App registered (or already present)
        '401':
          description: Missing or invalid admin key
        '422':
          description: Invalid appId or origins

  /health:
    get:
      summary: Server health check
      responses:
        '200':
          description: Healthy
        '503':
          description: Unhealthy (DB disconnected)